First rought cut to split read-dict.c into two
